requirements.txt
astroid==2.4.1
Bootstrap-Flask==1.3.2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
Flask==1.1.2
Flask-WTF==0.14.3
gevent==20.5.2
greenlet==0.4.16
idna==2.9
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
Pillow==7.1.2
pycparser==2.20
pylint==2.5.2
python-dotenv==0.13.0
requests==2.23.0
six==1.15.0
toml==0.10.1
urllib3==1.25.9
websocket==0.2.1
websocket-client==0.57.0
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.1
zope.event==4.4
zope.interface==5.1.0